Remove ContentParser.java
