--- application-crawler/src/main/java/com/seo/parser/ContentParser.java
@@ -1,7 +1,0 @@
-package com.seo.parser;
-
-import com.seo.model.Content;
-
-public interface ContentParser {
-    Content parseContent(String htmlSource);
-}
