+++ application-crawler/src/main/java/com/seo/parser/Parser.java
@@ -1,8 +1,10 @@
 package com.seo.parser;
 
 import com.seo.model.ConnectionResponse;
+import com.seo.model.Content;
 import com.seo.model.WebPage;
 
 public interface Parser {
+    Content parseContent(String htmlSource);
     void parseLinks(WebPage webPage, ConnectionResponse connectionResponse);
 }
